+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Changelog
 
-## 1.0.0-SNAPSHOT
+## 1.0.0-BETA3 (2022-01-21)
+
+1. Remove mention in tweet
+
+## 1.0.0-BETA2 (2022-01-11)
+
+1. Add catch for invalid twitter token validation
+2. Update words
+
+## 1.0.0-BETA (2021-12-14)
 
 1. Add initial nextjs setup
 2. Integration discord auth
@@ -8,3 +17,4 @@
    - link accounts
 3. Adjust to client side verification
 4. Add twitter claim page
+5. Post a tweet to Claim, misc cleanup